test(lua): clearer names for the two http-lua test files
Rename the two HTTP Lua handler tests so each name reflects what it verifies:

- http_lua_handler_test.cpp (synchronous HttpScriptHandler unit test: .lua
  dispatch, .py -> 501, script-dir routing, path-traversal guard, ctx API)
  -> http_script_handler_test.cpp  (its subject is HttpScriptHandler, the
  language-agnostic dispatcher)

- http_lua_async_test.cpp (real HttpServer + concurrent requests proving .lua
  handlers run as coroutines concurrently on one IO thread)
  -> http_lua_handler_test.cpp

Update the file bodies (headers, tmp dirs, PASSED banners) and all build
wiring: Makefile, unittest/CMakeLists.txt, scripts/unittest.sh. Both tests
build and pass under Make and CMake.
